코드 리뷰 반영, ConcatAdapter 적용
