+++ app/src/main/java/com/survivalcoding/todolist/data/datasource/ToDoInMemoryDataSource.kt
@@ -26,7 +26,6 @@ class ToDoInMemoryDataSource @Inject constructor() : ToDoLocalDataSource {
     }
 
     override fun addItem(newItem: ToDo) {
-        val tmpToDoList = _toDoList.value.toMutableList().apply { add(newItem) }
-        _toDoList.value = tmpToDoList
+        _toDoList.value = _toDoList.value.plus(newItem)
     }
 }
