+++ app/src/main/java/com/survivalcoding/todolist/presentation/todolist/ToDoListFragment.kt
@@ -9,13 +9,15 @@ import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
+import androidx.recyclerview.widget.ConcatAdapter
 import com.survivalcoding.todolist.databinding.FragmentToDoListBinding
 import com.survivalcoding.todolist.presentation.main.MainActivity
 import com.survivalcoding.todolist.presentation.main.MainViewModel
-import com.survivalcoding.todolist.presentation.main.ToDoListAdapter
+import com.survivalcoding.todolist.presentation.todolist.adapter.ToDoListAdapter
+import com.survivalcoding.todolist.presentation.todolist.adapter.ToDoListFooterAdapter
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
@@ -25,7 +27,7 @@ class ToDoListFragment : Fragment() {
 
     private var binding: FragmentToDoListBinding? = null
 
-    private val adapter: ToDoListAdapter by lazy {
+    private val toDoListBodyAdapter: ToDoListAdapter by lazy {
         ToDoListAdapter(
             onItemCheckedChanged = viewModel::changeDoneState,
             onDeleteButtonClick = viewModel::deleteToDo,
@@ -33,6 +35,12 @@ class ToDoListFragment : Fragment() {
                 (activity as MainActivity).navigateToCreateToDo(it)
             }
         )
+    }
+    private val toDoListFooterAdapter: ToDoListFooterAdapter by lazy {
+        ToDoListFooterAdapter()
+    }
+    private val toDoListConcatAdapter: ConcatAdapter by lazy {
+        ConcatAdapter(toDoListBodyAdapter, toDoListFooterAdapter)
     }
 
     override fun onCreateView(
@@ -45,7 +53,7 @@ class ToDoListFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        binding?.toDoListRecyclerView?.adapter = adapter
+        binding?.toDoListRecyclerView?.adapter = toDoListConcatAdapter
         binding?.createButton?.setOnClickListener {
             if (activity is MainActivity) {
                 (activity as MainActivity).navigateToCreateToDo()
@@ -56,7 +64,7 @@ class ToDoListFragment : Fragment() {
     }
 
     private fun collect() {
-        repeatOnStart { viewModel.toDoList.collect { adapter.submitList(it) } }
+        repeatOnStart { viewModel.toDoList.collectLatest { toDoListBodyAdapter.submitList(it) } }
     }
 
     private fun repeatOnStart(block: suspend CoroutineScope.() -> Unit) {
