+++ app/src/main/java/com/survivalcoding/todolist/presentation/todolist/adapter/ToDoListAdapter.kt
@@ -1,0 +1,18 @@
+package com.survivalcoding.todolist.presentation.todolist.adapter
+
+import android.view.ViewGroup
+import androidx.recyclerview.widget.ListAdapter
+import com.survivalcoding.todolist.domain.model.ToDo
+
+class ToDoListAdapter(
+    private val onItemCheckedChanged: (ToDo, Boolean) -> Unit,
+    private val onDeleteButtonClick: (ToDo) -> Unit,
+    private val onItemClick: (ToDo) -> Unit
+) : ListAdapter<ToDo, ToDoViewHolder>(ToDoDiffUtilItemCallback()) {
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) = ToDoViewHolder.from(parent)
+
+    override fun onBindViewHolder(holder: ToDoViewHolder, position: Int): Unit =
+        holder.bind(getItem(position), onItemCheckedChanged, onDeleteButtonClick, onItemClick)
+}
+
