+++ app/src/main/java/com/survivalcoding/todolist/presentation/todolist/adapter/ToDoViewHolder.kt
@@ -1,4 +1,4 @@
-package com.survivalcoding.todolist.presentation.main
+package com.survivalcoding.todolist.presentation.todolist.adapter
 
 import android.util.TypedValue
 import android.view.LayoutInflater
